ranobetest/requirements.txt
wheel 
django
django-rest-framework
psycopg2-binary
